MetaClasses: change types of fields in classes

+++ tryhard/MetaClasses.cs
@@ -101,18 +101,18 @@ namespace tryhard
 
     public class MaterialObject : BaseObject
     {
-        public long Weight { get; set; }
-        public long Volume { get; set; }
-        public long Cost { get; set; }
+        public float Weight { get; set; }
+        public float Volume { get; set; }
+        public float Cost { get; set; }
         public ObjectsStructure Structure { get; set; }
     }
 
     public class FieldParameters : BaseObject
     {
-        public long FluidOutput { get; set; }
+        public float FluidOutput { get; set; }
         public int HolesAmount { get; set; }
         public int OilQualityId { get; set; }
-        public long WaterInput { get; set; }
+        public float WaterInput { get; set; }
         public int ReserveDaysRvp { get; set; }
         public int ReserveDaysRtn { get; set; }
     }
@@ -126,15 +126,15 @@ namespace tryhard
 
     public class Pump : MaterialObject
     {
-        public long Input { get; set; }
-        public long Output { get; set; }
+        public float Input { get; set; }
+        public float Output { get; set; }
     }
 
     public class Pipe : MaterialObject
     {
         public int Diameter { get; set; }
         public string Material { get; set; }
-        public long Length { get; set; }
+        public float Length { get; set; }
     }
 
     public class Bg : MaterialObject
@@ -147,85 +147,85 @@ namespace tryhard
     public class ComplicatedObject : MaterialObject
     {
         public float PeopleDemand { get; set; }
-        public long ElectricityDemand { get; set; }
+        public float ElectricityDemand { get; set; }
     }
 
     public class Ngs : ComplicatedObject
     {
-        public long OilOutput { get; set; }
-        public long GasOutput { get; set; }
+        public float OilOutput { get; set; }
+        public float GasOutput { get; set; }
     }
 
     public class Sov : ComplicatedObject
     {
-        public long OilOutput { get; set; }
-        public long WaterOutput { get; set; }
+        public float OilOutput { get; set; }
+        public float WaterOutput { get; set; }
     }
 
     public class Nnpv : ComplicatedObject
     {
-        public long WaterInput { get; set; }
-        public long WaterOutput { get; set; }
+        public float WaterInput { get; set; }
+        public float WaterOutput { get; set; }
     }
 
     public class Fu : ComplicatedObject
     {
-        public long GasInput { get; set; }
+        public float GasInput { get; set; }
     }
 
     public class Rpv : ComplicatedObject
     {
-        public long WaterInput { get; set; }
-        public long WaterOutput { get; set; }
-        public long WaterVolume { get; set; }
+        public float WaterInput { get; set; }
+        public float WaterOutput { get; set; }
+        public float WaterVolume { get; set; }
     }
 
     public class Upn : ComplicatedObject
     {
-        public long FluidInput { get; set; }
-        public long OilOutput { get; set; }
-        public long WetGasOutput { get; set; }
-        public long WaterOutput { get; set; }
+        public float FluidInput { get; set; }
+        public float OilOutput { get; set; }
+        public float WetGasOutput { get; set; }
+        public float WaterOutput { get; set; }
     }
 
     public class Ukppv : ComplicatedObject
     {
-        public long WaterInput { get; set; }
-        public long WaterOutput { get; set; }
+        public float WaterInput { get; set; }
+        public float WaterOutput { get; set; }
     }
 
     public class Ukpg : ComplicatedObject
     {
-        public long WetGasInput { get; set; }
-        public long GasOutput { get; set; }
+        public float WetGasInput { get; set; }
+        public float GasOutput { get; set; }
     }
 
     public class Rtn : ComplicatedObject
     {
-        public long OilInput { get; set; }
-        public long OilOutput { get; set; }
-        public long OilVolume { get; set; }
+        public float OilInput { get; set; }
+        public float OilOutput { get; set; }
+        public float OilVolume { get; set; }
     }
 
     public class Dks : ComplicatedObject
     {
         public int PressureGradient { get; set; }
-        public long GasInput { get; set; }
-        public long GasOutput { get; set; }
+        public float GasInput { get; set; }
+        public float GasOutput { get; set; }
     }
 
     public class Rr : ComplicatedObject
     {
-        public long MaxReagentInput { get; set; }
+        public float MaxReagentInput { get; set; }
     }
 
     public class Dk : ComplicatedObject
     {
         public int HolesAmount { get; set; }
-        public long FluidInput { get; set; }
-        public long FluidOutput { get; set; }
-        public long WaterInput { get; set; }
-        public long ReagentInput { get; set; }
+        public float FluidInput { get; set; }
+        public float FluidOutput { get; set; }
+        public float WaterInput { get; set; }
+        public float ReagentInput { get; set; }
     }
 
     public class Complex: ComplicatedObject
@@ -235,21 +235,21 @@ namespace tryhard
 
     public class MiningComplex: Complex
     {
-        public long FluidOutput { get; set; }
+        public float FluidOutput { get; set; }
     }
 
     public class ProcessingComplex : Complex
     {
-        public long FluidInput { get; set; }
-        public long OilOutput { get; set; }
-        public long GasOutput { get; set; }
+        public float FluidInput { get; set; }
+        public float OilOutput { get; set; }
+        public float GasOutput { get; set; }
     }
 
     public class IntegratedComplex : Complex
     {
-        public long FluidInput { get; set; }
-        public long FluidOutput { get; set; }
-        public long OilOutput { get; set; }
-        public long GasOutput { get; set; }
+        public float FluidInput { get; set; }
+        public float FluidOutput { get; set; }
+        public float OilOutput { get; set; }
+        public float GasOutput { get; set; }
     }
 }
